% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_DEPRECATED.R
\name{apply_query}
\alias{apply_query}
\title{Apply query (inner)}
\usage{
apply_query(on, query, verbose = FALSE)
}
\arguments{
\item{verbose}{[\link{logical}]}
}
\value{

}
\description{
Apply query (inner)
}
