% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_DEPRECATED.R
\name{apply_query_iter}
\alias{apply_query_iter}
\title{Apply query (outer)}
\usage{
apply_query_iter(on, query, ops, verbose = FALSE)
}
\arguments{
\item{verbose}{[\link{logical}]}
}
\value{

}
\description{
Apply query (outer)
}
