% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{conf_auto_load}
\alias{conf_auto_load}
\title{Load project configs \lifecycle{experimental}}
\usage{
conf_auto_load()
}
\value{
[\link{list}]
}
\description{
Put this in your custom \code{.onAttach()} function to load your project's configs
based on \code{.yml} files that typically should live in your project's root
directory (check via \code{\link[here:here]{here::here()}}).
}
\details{
TODO-20191115-3: Explain env var \code{CONFX_AUTO_LOAD}
}
