% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{conf_auto_load_internal}
\alias{conf_auto_load_internal}
\title{Load internal package configs \lifecycle{experimental}}
\usage{
conf_auto_load_internal(pkgname)
}
\arguments{
\item{pkgname}{[\code{\link[=character]{character()}}] See \link{?.onAttach}}
}
\value{
[\link{list}]
}
\description{
Put this in your custom \code{.onAttach()} function to load your package's
internal configs based on \code{.yml} files that typically should live in your
\code{./inst} directory (which becomes the root package directory once the package
is attached).
}
\details{
TODO-20191115-2: Explain env var \code{CONFX_AUTO_LOAD_INTERNAL}
}
