% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{conf_get}
\alias{conf_get}
\title{Get config \lifecycle{experimental}}
\usage{
conf_get(value = character(), from = "config.yml",
  dir_from = here::here(), sep = "/", inheritance_handling = TRUE)
}
\arguments{
\item{value}{[[character]]}

\item{from}{[[character]]}

\item{dir_from}{[[character]]}

\item{sep}{[[character]]}

\item{inheritance_handling}{[[logical]]}
}
\value{
[[list]]
}
\description{
TODO-20191026-1: Write doc for `conf_get()`
}
