% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{conf_get_reference_id_inherited}
\alias{conf_get_reference_id_inherited}
\title{Reference ID for \code{inherits} YAML entity}
\usage{
conf_get_reference_id_inherited()
}
\value{

}
\description{
Reference ID for \code{inherits} YAML entity
}
