% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conf_handle_inherited}
\alias{conf_handle_inherited}
\title{Title}
\usage{
conf_handle_inherited(configs, from, dir_from, name = "inherits")
}
\arguments{
\item{name}{}
}
\description{
TODO-20191026-10: Write doc for `conf_handle_inherited()`
}
