% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{conf_handle_reference}
\alias{conf_handle_reference}
\title{Title}
\usage{
conf_handle_reference(value, from)
}
\arguments{
\item{value}{[\link{value}]}

\item{from}{[\link{value}]}
}
\value{
[\link{value}]
}
\description{
Title
}
