% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{conf_handle_reference_file}
\alias{conf_handle_reference_file}
\title{Handle file references}
\usage{
conf_handle_reference_file(value, from)
}
\arguments{
\item{value}{[\link{value}]}

\item{from}{[\link{value}]}
}
\value{
[\link{value}]
}
\description{
Handle file references
}
