% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{conf_handle_reference_inherited}
\alias{conf_handle_reference_inherited}
\title{Handle references to inherited configs}
\usage{
conf_handle_reference_inherited(
  configs,
  from,
  dir_from = Sys.getenv("R_CONFIG_DIR", getwd()),
  drop_ref_link = TRUE
)
}
\arguments{
\item{configs}{[\link{list}]}

\item{from}{[\link{character}]}

\item{dir_from}{[\link{character}]}

\item{drop_ref_link}{[\link{logical}]}
}
\value{
[\link{list} or \link{character}]
}
\description{
TODO-20191026-10: Write doc for \code{conf_handle_reference_inherited()}
}
