% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{conf_handle_reference_scope}
\alias{conf_handle_reference_scope}
\title{Handle reference scope}
\usage{
conf_handle_reference_scope(from, from_this)
}
\arguments{
\item{from_this}{}
}
\value{

}
\description{
Handle reference scope
}
