% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conf_has_inherited}
\alias{conf_has_inherited}
\title{Title}
\usage{
conf_has_inherited(configs, name = "inherits")
}
\arguments{
\item{configs}{[\link{list}]}

\item{name}{[\link{character}]}
}
\value{
[\link{logical}]
}
\description{
TODO-20191026-7: Write doc for \code{conf_has_inherited()}
}
