% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{conf_has_reference}
\alias{conf_has_reference}
\title{Check for existence of \code{yaml} file reference}
\usage{
conf_has_reference(value)
}
\arguments{
\item{value}{[\link{character}]}
}
\value{
[\link{logical}]
}
\description{
Check for existence of \code{yaml} file reference
}
