% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{conf_has_reference_inherits}
\alias{conf_has_reference_inherits}
\title{Check for existence of \code{inherits} reference}
\usage{
conf_has_reference_inherits(configs, name = "inherits")
}
\arguments{
\item{configs}{[\link{list}]}

\item{name}{[\link{character}]}
}
\value{
[\link{logical}]
}
\description{
Check for existence of \code{inherits} file reference
}
\details{
TODO-20191026-7: Write doc for \code{conf_has_reference_inherits()}
}
