% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{conf_has_reference_json}
\alias{conf_has_reference_json}
\title{Check for existence of \verb{$ref} reference}
\usage{
conf_has_reference_json(configs)
}
\arguments{
\item{configs}{[\link{list}]}
}
\value{
[\link{logical}]
}
\description{
Check for existence of \verb{$ref} reference
}
