% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conf_index_recursively}
\alias{conf_index_recursively}
\title{Title}
\usage{
conf_index_recursively(lst, el, .el_trace = el, .level_trace = 1)
}
\arguments{
\item{lst}{[[list]]}

\item{el}{[[character]]}

\item{.el_trace}{[[character]]}

\item{.level_trace}{[[character]]}
}
\value{
[[character]] or [[list]]
}
\description{
TODO-20191026-5: Write doc for `conf_index_recursively()`
}
