% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{conf_load}
\alias{conf_load}
\title{Load configs into options \lifecycle{experimental}}
\usage{
conf_load(dir = here::here())
}
\arguments{
\item{dir}{[[character]]}
}
\value{
[[list]]
}
\description{
TODO-20191026-2: Write doc for `conf_load()`
}
