% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{conf_load_from_dir}
\alias{conf_load_from_dir}
\title{Load all configs in a directory into options \lifecycle{experimental}}
\usage{
conf_load_from_dir(
  dir = getwd(),
  regexp_valid_files = SYS_VALID_CONF_EXTENSIONS(as_regexp = TRUE),
  regexp_disregarded_files = SYS_DISREGARDED_CONF_FILES(as_regexp = TRUE)
)
}
\arguments{
\item{dir}{[\link{character}]}
}
\value{
[\link{list}]
}
\description{
\itemize{
\item Lists all files in a given \code{dir}
\item Greps for files with extension as defined/returned by
\code{.SYS_VALID_CONF_EXTENSIONS()} while disregarding extensions as
defined/returned by `.SYS_DISREGARDED_FILE_EXTENSIONS()``
\item Maps \code{conf_load_from_file()} to those grep results
}
}
\details{
TODO-20191026-2: Write doc for \code{conf_load_from_dir()}
}
\seealso{
\code{\link[conf:conf_load_from_file]{conf::conf_load_from_file()}}
}
