% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{conf_load_from_file}
\alias{conf_load_from_file}
\title{Load configs from file}
\usage{
conf_load_from_file(file, sep_opt_name = "_", verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Load configs from file
}
