% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{conf_merge}
\alias{conf_merge}
\title{Assign configs to local variables \lifecycle{experimental}}
\usage{
conf_merge(config_x, config_y)
}
\arguments{
\item{config_y}{}
}
\description{
TODO-20191026-4: Write doc for `conf_merge()`
}
