% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conf_merge_inherited}
\alias{conf_merge_inherited}
\title{Title}
\usage{
conf_merge_inherited(configs_inherited, configs)
}
\arguments{
\item{configs_inherited}{[[list]]}

\item{configs}{[[list]]}
}
\value{
[[list]]
}
\description{
TODO-20191026-9: Write doc for `conf_merge_inherited()`
}
