% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conf_merge_lists}
\alias{conf_merge_lists}
\title{Title}
\usage{
conf_merge_lists(base_list, overlay_list, recursive = TRUE,
  is_vector = FALSE)
}
\arguments{
\item{base_list}{[\link{list}]}

\item{overlay_list}{[\link{list}]}

\item{recursive}{[\link{logical}]}

\item{is_vector}{[\link{logical}]}
}
\value{
[\link{list}]
}
\description{
TODO-20191026-6: Write doc for \code{conf_merge_lists()}
}
