% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{conf_merge_referenced}
\alias{conf_merge_referenced}
\title{Title}
\usage{
conf_merge_referenced(configs_inherited, configs)
}
\arguments{
\item{configs_inherited}{[\link{list}]}

\item{configs}{[\link{list}]}
}
\value{
[\link{list}]
}
\description{
TODO-20191026-9: Write doc for \code{conf_merge_referenced()}
}
