% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conf_resolve_config_reference}
\alias{conf_resolve_config_reference}
\title{Title}
\usage{
conf_resolve_config_reference(value)
}
\arguments{
\item{value}{[\link{character}]}
}
\value{
[\link{character}]
}
\description{
Title
}
