% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conf_resolve_inherited}
\alias{conf_resolve_inherited}
\title{Title}
\usage{
conf_resolve_inherited(value_reference, from, dir_from)
}
\arguments{
\item{value_reference}{[[list] or [character]]}

\item{from}{[[character]]}

\item{dir_from}{[[character]]}
}
\value{
[[list] or [character]]
}
\description{
TODO-20191026-8: Write doc for `conf_resolve_inherited()`
}
