% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{conf_resolve_reference_file}
\alias{conf_resolve_reference_file}
\title{Resolve file reference}
\usage{
conf_resolve_reference_file(value)
}
\arguments{
\item{value}{[\link{character}]}
}
\value{
[\link{character}]
}
\description{
Resolve file reference
}
