% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{conf_resolve_reference_json}
\alias{conf_resolve_reference_json}
\title{Resolve JSON reference}
\usage{
conf_resolve_reference_json(path_reference, from, dir_from)
}
\arguments{
\item{path_reference}{[\link{list} or \link{character}]}

\item{from}{[\link{character}]}

\item{dir_from}{[\link{character}]}
}
\value{
[\link{list} or \link{character}]
}
\description{
TODO-20191026-8: Write doc for \code{conf_resolve_reference_json()}
}
