% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{conf_resolve_reference_json_}
\alias{conf_resolve_reference_json_}
\title{Resolve JSON reference (inner)}
\usage{
conf_resolve_reference_json_(uri)
}
\arguments{
\item{uri}{[\link{character}]}
}
\value{

}
\description{
Resolve JSON reference (inner)
}
