% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{eval_query_call_two_args}
\alias{eval_query_call_two_args}
\title{Eval query call (two args)}
\usage{
eval_query_call_two_args(
  .op,
  .query,
  on,
  recursive_fn = apply_query,
  verbose = FALSE
)
}
\arguments{
\item{recursive_fn}{}

\item{verbose}{[\link{logical}]}
}
\value{

}
\description{
Eval query call (two args)
}
