% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{handle_conf_query}
\alias{handle_conf_query}
\title{Handle config query}
\usage{
handle_conf_query(x, query, op, verbose = FALSE)
}
\arguments{
\item{verbose}{[\link{logical}]}
}
\value{

}
\description{
Currently, this is a list-oriented query
}
