% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_DEPRECATED.R
\name{handle_extracted_operators}
\alias{handle_extracted_operators}
\title{Handle extracted operator}
\usage{
handle_extracted_operators(op)
}
\arguments{
\item{op}{}
}
\value{

}
\description{
TODO-20201229-1121: Handle multiple extracted operators
}
