% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_DEPRECATED.R
\name{handle_query_result}
\alias{handle_query_result}
\title{Handle query result}
\usage{
handle_query_result(result, ops, handler_fn = all)
}
\arguments{
\item{resolve_fn}{}
}
\value{

}
\description{
Handle vector results and apply recursive chain logic.
}
