% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_DEPRECATED.R
\name{handle_query_result_vectorized}
\alias{handle_query_result_vectorized}
\title{Handle vectorized query result}
\usage{
handle_query_result_vectorized(result, ops, fn = all)
}
\arguments{
\item{fn}{}
}
\value{

}
\description{
Handle vectorized query result
}
