% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_DEPRECATED.R
\name{handle_query_vectorized}
\alias{handle_query_vectorized}
\title{Handle vectorized query}
\usage{
handle_query_vectorized(query)
}
\arguments{
\item{query}{}
}
\value{

}
\description{
Handle vectorized query
}
