% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{handle_query_vectorized_2}
\alias{handle_query_vectorized_2}
\title{Handle vectorized query}
\usage{
handle_query_vectorized_2(call, fn = "all")
}
\arguments{
\item{fn}{}
}
\value{

}
\description{
Handle vectorized query
}
