% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_DEPRECATED.R
\name{parse_query_rhs}
\alias{parse_query_rhs}
\title{Parse right-hand side of query}
\usage{
parse_query_rhs(query)
}
\arguments{
\item{query}{}
}
\value{

}
\description{
Takes care of transforming a vector notation into an actual vector for the
second query subitem
}
