% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexp.R
\name{regexp__file_extensions_as_regexp}
\alias{regexp__file_extensions_as_regexp}
\title{File extensions as regexp}
\usage{
regexp__file_extensions_as_regexp(
  x,
  prefix = ".*",
  suffix = "$",
  wrap_in_parentheses = TRUE,
  include_uppercase = TRUE,
  escape_dots = TRUE
)
}
\arguments{
\item{escape_dots}{}
}
\value{

}
\description{
File extensions as regexp
}
