% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_recursively}
\alias{subset_recursively}
\title{Subset object recursively}
\usage{
subset_recursively(
  x,
  index,
  leaf_as_list = FALSE,
  .index_trace = index,
  .level_trace = 1,
  verbose = FALSE
)
}
\arguments{
\item{x}{[\link{list}]}

\item{index}{[\link{character}]}

\item{leaf_as_list}{[\link{logical}]}

\item{.index_trace}{[\link{character}]}

\item{.level_trace}{[\link{character}]}

\item{verbose}{[\link{logical}]}
}
\value{
[\link{character}] or [\link{list}]
}
\description{
TODO-20191026-5: Write doc for \code{subset_recursively()}
}
