% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.R
\name{SYS_DISREGARDED_CONF_FILES}
\alias{SYS_DISREGARDED_CONF_FILES}
\title{Global variable for disregarded conf files}
\usage{
SYS_DISREGARDED_CONF_FILES(as_regexp = FALSE)
}
\arguments{
\item{as_regexp}{}
}
\value{

}
\description{
Global variable for disregarded conf files
}
