% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.R
\name{SYS_VALID_CONF_EXTENSIONS}
\alias{SYS_VALID_CONF_EXTENSIONS}
\title{Global variable for valid conf extensions}
\usage{
SYS_VALID_CONF_EXTENSIONS(as_regexp = FALSE)
}
\arguments{
\item{as_regexp}{}
}
\value{

}
\description{
Global variable for valid conf extensions
}
