% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_DEPRECATED.R
\name{tokenize_query_items}
\alias{tokenize_query_items}
\title{Tokenize query items}
\usage{
tokenize_query_items(query, ops = valid_operators_logical())
}
\arguments{
\item{ops}{}
}
\value{

}
\description{
Tokenize query items
}
