% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_operators_logical}
\alias{valid_operators_logical}
\title{Valid logical operators}
\usage{
valid_operators_logical(
  x = character(),
  reverse = FALSE,
  flip = FALSE,
  strict = TRUE,
  as_regexp = FALSE
)
}
\arguments{
\item{as_regexp}{}
}
\value{

}
\description{
Valid logical operators
}
