% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_operators_query_chain}
\alias{valid_operators_query_chain}
\title{Valid query chain operators}
\usage{
valid_operators_query_chain(
  x = character(),
  reverse = FALSE,
  flip = FALSE,
  strict = TRUE
)
}
\arguments{
\item{strict}{}
}
\value{

}
\description{
Valid query chain operators
}
