% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{validate_json_reference}
\alias{validate_json_reference}
\title{Validate JSON references}
\usage{
validate_json_reference(uri, regexp = ".*#/.*(/.*)?")
}
\arguments{
\item{uri}{[\link{character}]}

\item{regexp}{[\link{character}]}
}
\value{

}
\description{
Validate JSON references
}
