% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_DEPRECATED.R
\name{wrap_query_with_fn}
\alias{wrap_query_with_fn}
\title{Wrap query with additional function}
\usage{
wrap_query_with_fn(query, op, fn = "all")
}
\arguments{
\item{fn}{}
}
\value{

}
\description{
Wrap query with additional function
}
