% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{conf_assign}
\alias{conf_assign}
\title{Assign configs to local variables \lifecycle{experimental}}
\usage{
conf_assign(config_list, env = rlang::caller_env(),
  from = "config.yml", dir_from = here::here(), sep = "/")
}
\arguments{
\item{config_list}{[[list]]}

\item{env}{[[environment]]}

\item{from}{[[character]]}

\item{dir_from}{[[character]]}

\item{sep}{[[character]]}
}
\value{
[[list]]
}
\description{
TODO-20191026-3: Write doc for `conf_assign()`
}
