% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{conf_assign}
\alias{conf_assign}
\title{Assign configs to local variables \lifecycle{experimental}}
\usage{
conf_assign(config_list, env = rlang::caller_env(),
  from = "config.yml", dir_from = here::here(), sep = "/")
}
\arguments{
\item{config_list}{[\link{list}]}

\item{env}{[\link{environment}]}

\item{from}{[\link{character}]}

\item{dir_from}{[\link{character}]}

\item{sep}{[\link{character}]}
}
\value{
[\link{list}]
}
\description{
TODO-20191026-3: Write doc for \code{conf_assign()}
}
