% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{conf_get}
\alias{conf_get}
\title{Get config value from in-memory cache or from file  \lifecycle{experimental}}
\usage{
conf_get(
  value = character(),
  from = "config.yml",
  dir_from = Sys.getenv("R_CONFIG_DIR", getwd()),
  config = Sys.getenv("R_CONFIG_ACTIVE", "default"),
  use_parent = TRUE,
  sep = "/",
  sep_opt_name = "_",
  resolve_references = TRUE,
  force_from_file = FALSE,
  leaf_as_list = FALSE
)
}
\arguments{
\item{value}{[\link{character}]}

\item{from}{[\link{character}]}

\item{dir_from}{[\link{character}]}

\item{config}{[\link{character}]}

\item{use_parent}{[\link{logical}]}

\item{sep}{[\link{character}]}

\item{sep_opt_name}{[\link{character}]}

\item{resolve_references}{[\link{logical}]}

\item{force_from_file}{[\link{logical}]}

\item{leaf_as_list}{[\link{logical}]}
}
\value{
[\link{list} or \link{character}]
}
\description{
TODO-20191026-1: Write doc for \code{conf_get()}
}
