% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{conf_get}
\alias{conf_get}
\title{Get config \lifecycle{experimental}}
\usage{
conf_get(value = character(), from = "config.yml",
  dir_from = here::here(), sep = "/", sep_opt_name = "_",
  inheritance_handling = TRUE, force_from_file = FALSE,
  leaf_as_list = FALSE)
}
\arguments{
\item{value}{[\link{character}]}

\item{from}{[\link{character}]}

\item{dir_from}{[\link{character}]}

\item{sep}{[\link{character}]}

\item{sep_opt_name}{[\link{character}]}

\item{inheritance_handling}{[\link{logical}]}

\item{force_from_file}{[\link{logical}]}

\item{leaf_as_list}{[\link{logical}]}
}
\value{
[\link{list} or \link{character}]
}
\description{
TODO-20191026-1: Write doc for \code{conf_get()}
}
