% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conf_handle_config_reference}
\alias{conf_handle_config_reference}
\title{Title}
\usage{
conf_handle_config_reference(value, from)
}
\arguments{
\item{value}{[[value]]}

\item{from}{[[value]]}
}
\value{
[[value]]
}
\description{
Title
}
