% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conf_handle_inherited}
\alias{conf_handle_inherited}
\title{Title}
\usage{
conf_handle_inherited(
  configs,
  from,
  dir_from = here::here(),
  name = "inherits"
)
}
\arguments{
\item{configs}{[\link{list}]}

\item{from}{[\link{character}]}

\item{dir_from}{[\link{character}]}

\item{name}{[\link{character}]}
}
\value{
[\link{list} or \link{character}]
}
\description{
TODO-20191026-10: Write doc for \code{conf_handle_inherited()}
}
