% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conf_index_recursively}
\alias{conf_index_recursively}
\title{Title}
\usage{
conf_index_recursively(lst, el, leaf_as_list = FALSE, .el_trace = el,
  .level_trace = 1)
}
\arguments{
\item{lst}{[\link{list}]}

\item{el}{[\link{character}]}

\item{leaf_as_list}{[\link{logical}]}

\item{.el_trace}{[\link{character}]}

\item{.level_trace}{[\link{character}]}
}
\value{
[\link{character}] or [\link{list}]
}
\description{
TODO-20191026-5: Write doc for \code{conf_index_recursively()}
}
