% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{conf_load}
\alias{conf_load}
\title{Load configs into options \lifecycle{experimental}}
\usage{
conf_load(dir = here::here(), pattern_disregard = "^(_|\\\\.|codecov|travis)")
}
\arguments{
\item{dir}{[\link{character}]}
}
\value{
[\link{list}]
}
\description{
TODO-20191026-2: Write doc for \code{conf_load()}
}
