% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{conf_merge}
\alias{conf_merge}
\title{Assign configs to local variables \lifecycle{experimental}}
\usage{
conf_merge(config_x, config_y)
}
\arguments{
\item{config_x}{[\link{list}]}

\item{config_y}{[\link{list}]}
}
\value{
[\link{list}]
}
\description{
TODO-20191026-4: Write doc for \code{conf_merge()}
}
