% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conf_merge_lists}
\alias{conf_merge_lists}
\title{Title}
\usage{
conf_merge_lists(base_list, overlay_list, recursive = TRUE,
  is_vector = FALSE)
}
\arguments{
\item{is_vector}{}
}
\description{
TODO-20191026-6: Write doc for `conf_merge_lists()`
}
