% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conf_resolve_inherited}
\alias{conf_resolve_inherited}
\title{Title}
\usage{
conf_resolve_inherited(value_reference, from, dir_from)
}
\arguments{
\item{dir_from}{}
}
\description{
TODO-20191026-8: Write doc for `conf_resolve_inherited()`
}
